Refactor dark mode (context not needed)
