+++ src/views/public-blog/components/DarkModeToggle.jsx
@@ -1,9 +1,32 @@
 import { HiOutlineMoon, HiOutlineSun } from 'react-icons/hi2';
-import { DarkModeContext } from '../context/DarkModeContext';
-import { useContext } from 'react';
+import { useEffect, useState } from 'react';
 
 function DarkModeToggle() {
-  const { isDarkMode, toggleDarkMode } = useContext(DarkModeContext);
+  const [isDarkMode, setIsDarkMode] = useState(function () {
+    const storedValue = localStorage.getItem('dark-mode');
+    return storedValue ? JSON.parse(storedValue) : false;
+  });
+
+  function toggleDarkMode() {
+    setIsDarkMode(!isDarkMode);
+  }
+
+  useEffect(
+    function () {
+      isDarkMode
+        ? document.documentElement.classList.add('dark')
+        : document.documentElement.classList.remove('dark');
+    },
+    [isDarkMode],
+  );
+
+  useEffect(
+    function () {
+      localStorage.setItem('dark-mode', JSON.stringify(isDarkMode));
+    },
+    [isDarkMode],
+  );
+
   return (
     <button onClick={toggleDarkMode}>
       {isDarkMode ? (
