+++ tailwind.config.js
@@ -19,5 +19,5 @@ module.exports = {
     },
   },
   plugins: [require('tailwind-scrollbar')],
-  darkMode: 'class',
+  darkMode: 'selector',
 };
